Move encodeURI and parseAuthHeader into http package, alphabetize lib versions
