+++ cli/shared/src/main/scala/com/lightbend/rp/reactivecli/docker/DockerEngine.scala
@@ -20,7 +20,6 @@ import argonaut._
 import com.lightbend.rp.reactivecli.concurrent._
 import com.lightbend.rp.reactivecli.files._
 import com.lightbend.rp.reactivecli.http._
-import com.lightbend.rp.reactivecli.Platform
 import scala.concurrent.{ ExecutionContext, Future }
 import slogging._
 
@@ -54,7 +53,7 @@ object DockerEngine extends LazyLogging {
       case Some(host) if host.startsWith("tcp://") =>
         val verify = env.get("DOCKER_TLS_VERIFY").contains("1")
         val protocol = if (verify) "https" else "http"
-        val url = Platform.encodeURI(s"$protocol://${host.replaceFirst("tcp://", "")}/images/$uri/json")
+        val url = encodeURI(s"$protocol://${host.replaceFirst("tcp://", "")}/images/$uri/json")
 
         logger.debug("Attempting to pull config from Engine, {}", url)
